requirements.txt
streamlit==1.45.0 
anthropic 
python-dotenv 
stripe 
supabase